SDK-962: Use configured Joomla folder
